statementData生成処理をファイルとして独立させた
