+++ src/1-theater/types.ts
@@ -1,0 +1,31 @@
+export type PerformanceRecord = {
+  playID: string;
+  audience: number;
+};
+
+export type InvoiceRecord = {
+  customer: string;
+  performances: PerformanceRecord[];
+};
+
+export type Play = {
+  name: string;
+  type: string;
+};
+
+export type PlaysRecord = { [playID: string]: Play };
+
+export type Performance = {
+  playID: string;
+  play: Play;
+  audience: number;
+  amount: number;
+  volumeCredits: number;
+};
+
+export type StatementData = {
+  customer: string;
+  performances: Performance[];
+  totalAmount: number;
+  totalVolumeCredits: number;
+};
